ENG-03: clear mypy errors in 7 modules (batch 1)
Behaviour-preserving type fixes:
- _analyses/model_selection: annotate current_terms once (no-redef)
- _analyses/box_cox, analysis: np.asarray(...) instead of .values.astype(float)
- monitoring/metrics: float()-wrap center/spread so _safe_ratio gets float
- experiments/augment: .to_numpy() instead of .values for the *= operand
- experiments/visualization/api: cast df.to_dict(records) to list[dict[str, Any]]
- experiments/strategy/engine: cast the validated detail_level to its Literal

mypy clean on all 7; ruff clean; affected tests pass.

+++ src/process_improve/monitoring/metrics.py
@@ -84,11 +84,11 @@ def calculate_cpk(  # noqa: C901
     metric_upper = Cpk_upper_spec - df[which_column]
 
     if trim_percentile > 0:
-        center_lower, center_upper = metric_lower.median(), metric_upper.median()
-        spread_lower, spread_upper = Sn(metric_lower), Sn(metric_upper)
+        center_lower, center_upper = float(metric_lower.median()), float(metric_upper.median())
+        spread_lower, spread_upper = float(Sn(metric_lower)), float(Sn(metric_upper))
     else:
-        center_lower, center_upper = metric_lower.mean(), metric_upper.mean()
-        spread_lower, spread_upper = metric_lower.std(), metric_upper.std()
+        center_lower, center_upper = float(metric_lower.mean()), float(metric_upper.mean())
+        spread_lower, spread_upper = float(metric_lower.std()), float(metric_upper.std())
 
     # A column with no spread (constant data, or only one non-NaN value)
     # makes Cpk undefined: a bare division would silently yield inf / NaN.
